refactor: Quantity Class common 패키지로 이동
