+++ src/main/java/kitchenpos/common/Price.java
@@ -6,7 +6,6 @@ import java.math.BigDecimal;
 import java.util.Objects;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
-import kitchenpos.menu.domain.Quantity;
 
 @Embeddable
 public class Price implements Comparable<Price> {
