+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -9,6 +9,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import kitchenpos.common.Quantity;
 
 @Entity
 public class MenuProduct {
