+++ src/main/java/kitchenpos/menu/domain/MenuProducts.java
@@ -1,13 +1,11 @@
 package kitchenpos.menu.domain;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.OneToMany;
-import kitchenpos.common.Price;
 
 @Embeddable
 public class MenuProducts {
