+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -9,7 +9,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import kitchenpos.menu.domain.Quantity;
+import kitchenpos.common.Quantity;
 
 @Entity
 public class OrderLineItem {
