+++ src/main/java/kitchenpos/product/domain/Product.java
@@ -9,7 +9,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import kitchenpos.common.Price;
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.Quantity;
+import kitchenpos.common.Quantity;
 
 @Entity
 public class Product {
