+++ src/test/java/kitchenpos/menu/domain/MenuProductsTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.Arrays;
 import kitchenpos.common.Price;
 
+import kitchenpos.common.Quantity;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductName;
 import org.junit.jupiter.api.BeforeEach;
